Add continuous deployment

+++ deploy.ps1
@@ -4,75 +4,75 @@ param (
 )
 
 $environment = Get-Content 'developers.json' | Out-String | ConvertFrom-Json
-
 $config = $environment.$configName
+$developerPrefix = $config.Prefix
 
 $stackName = "pecuniary-transactionquery-stack"
 
-$developerPrefix = $config.Prefix
-
-Write-Host "`Parameters from " -NoNewline
-Write-Host "developers.json:`n" -ForegroundColor Cyan
-Write-Host "`tdeveloperPrefix: `t`t $developerPrefix" -ForegroundColor Yellow
-
 $sourceFile = "samTemplate.yaml"
-$localFileName = "$sourceFile.local"
-Write-Host "`nCreating/updating $localFileName based on $sourceFile..."
-
-Copy-Item samTemplate.yaml $localFileName
+$localSourceFile = "$sourceFile.local"
+Write-Host "`nCreating/updating $localSourceFile based on $sourceFile..."
+Copy-Item samTemplate.yaml $localSourceFile
 
 if ($config.Prefix)
 {  
     Write-Host "`n`tDeveloper config selected" -ForegroundColor Yellow
+    Write-Host "`Parameters from " -NoNewline
+    Write-Host "developers.json:`n" -ForegroundColor Cyan
+    Write-Host "`tdeveloperPrefix: `t`t $developerPrefix" -ForegroundColor Yellow
 
-    $stackName = $developerPrefix-$stackName
+    $stackName = $developerPrefix + "-" + $stackName
 
-    (Get-Content $localFileName) `
+    (Get-Content $localSourceFile) `
         -replace 'pecuniary-', "$developerPrefix-pecuniary-" `
         -replace 'Name: pecuniary', "Name: $developerPrefix-pecuniary" |
-    Out-File $localFileName -Encoding utf8
-}
+    Out-File $localSourceFile -Encoding utf8
 
-Write-Host "`nDone! $localFileName updated. Please use this file when deploying to our own AWS stack.`n"
+    Write-Host "`nDone! $localSourceFile updated. Please use this file when deploying to our own AWS stack.`n"
 
-if ($config.Prefix) 
-{ 
     Write-Host "Press [enter] to continue deploying stack to AWS (Ctrl+C to exit)" -NoNewline -ForegroundColor Green
     Read-Host
+
+    Write-Host "`n`nRestoring projects..." -ForegroundColor Cyan
+
+    dotnet restore
+
+    Write-Host "`n`nBuilding projects..." -ForegroundColor Cyan
+
+    dotnet publish -c Release
 }
 
-$samTemplate = 'samTemplate.yaml.local'
-
-Write-Host "`n`nPrebuild:"
-
-dotnet restore Pecuniary.Transaction.Events/Pecuniary.Transaction.Events.csproj
-dotnet restore Pecuniary.Transaction.Query/Pecuniary.Transaction.Query.csproj
-
-Write-Host "`n`nBuild:"
-
-dotnet publish Pecuniary.Transaction.Events/Pecuniary.Transaction.Events.csproj
-dotnet publish Pecuniary.Transaction.Query/Pecuniary.Transaction.Query.csproj
-
-Write-Host "`n`nDeploy:"
+Write-Host "`n`nDeploying stack $stackName..." -ForegroundColor Cyan
 
 dotnet-lambda deploy-serverless `
     --stack-name $stackName `
-    --template $samTemplate `
+    --template $localSourceFile `
     --region us-west-2 `
     --s3-bucket pecuniary-deployment-artifacts
 
+# Handle deploy errors
+if ($lastexitcode -ne 0) {
+    throw "Error deploying" + $stackName
+}
+
 # Get the API Gateway Base URL
-$stack = aws cloudformation describe-stacks --stack-name $stackName | ConvertFrom-Json
+$stack = aws cloudformation describe-stacks --stack-name $stackName --region us-west-2 | ConvertFrom-Json
 $outputKey = $stack.Stacks.Outputs.OutputKey.IndexOf("PecuniaryApiGatewayBaseUrl")
 $apiGatewayBaseUrl = $stack.Stacks.Outputs[$outputKey].OutputValue
 
 # Add scopes
-Write-Host "`n`Adding Scopes to $apiGatewayBaseUrl"
+Write-Host "`n`Adding OAuth scopes..."
 aws lambda invoke `
     --function-name "pecuniary-AddScopes" `
     --payload """{ """"ApiGatewayBaseUrl"""": """"$apiGatewayBaseUrl"""" }""" `
+    --region us-west-2 `
     outfile.json
+
+# Handle add scope errors
+if ($lastexitcode -ne 0) {
+    throw "Error adding OAuth scopes"
+}
 Remove-Item outfile.json
 
-Write-Host "`n`n YOUR STACK NAME IS:   $stackName   `n`n"
- 
+Write-Host "`n`n YOUR STACK NAME IS:   " -NoNewLine
+Write-Host "$stackName   `n`n" -ForegroundColor Cyan
